Remove usage of a local runtime cache to using redis for tree traversal
We had been using a weirdo __global__ cache for the asset tree traveral (quick lookup of nodes). This sucks because it means we don't persist the db over executions. Well we sorta still don't but now at least we are using redis db for the lookup and storage. This means nodes are stored and persist ... only on my local machine.

Also the way things are right now, since every run is creating a new PID for every node (randomized) so we don't really go and access the same nodes the next time the code runs lol. Will fix this in the upcoming commits somehow