Adding support for image storage and retrieval
Creating another DataBase called ImageStore. This is specific to images and uses sqlite to store images as binary blobs. A simple schema with just two columns, first is the image handle which is technically the PID of the Image node. Second is the blob data itself which is essentially a binary string representing the image.

Since Image is a node it is also stored in Node DB but note how the image blob itself is not stored in Node DB (since it is supposed to be a fast lookup cache).

All images are attached to Users through special edges. These edges are not traversed as a part of asset trees!

Lots of corner cases have been created and had to be debugged to make everythin work here, maybe we can use a design refresh at some point.