Ingest full image directly
Allows a User node to directly ingest a full image into it's asset tree. This reads the image, uses the llm engine to identify all objects in the image and then ingests them (as Assets) into the asset tree.

NOTE!!!!!!!
We are not currently linking images to asset nodes, i.e there is no edge from Asset => Image. We will deliberate if this is needed at a future time.