Add ability to add descriptors through UI
A logged-in user can now add descriptors (Asset Categories) to their asset tree manually throught the UI.

If the descriptor is added successfully the UI will say so and display the ID of added descriptor. If not, the UI will say it failed.

TODO: Make the UI prettier.
TODO: Add a "loading" wheel after clicking the Add Asset Category button and before results load.

Tested everything through UI, it works.